xrp, xrp_viewer: Add more query feedback.
Now informs the user if its too soon to query for changes, or if the
target does not appear to have an RP addon, or if the target is offline.
First query (if not already queried for tooltips, etc.) can still show
no response, if target is online and does not have an RP addon --
nothing to be done about that.