xrp_editor: Clean up field checking.
This cleans up the field checking to be a little bit leaner on CPU time
when possible (it was, and still can, cause momentary framerate hiccups
when checking all fields -- but this makes that less common). It also
fills in fallback values for name (toon name), race (toon race), and
class (toon class) which are sent automatically if those fields are
empty.