Add alternate race display names.
Toggles are added that will swap Blood Elf, Night Elf, Tauren, and
Undead to display as Sin'dorei, Kaldorei, Shu'halo, and Forsaken,
respectively. These are primarily only on the client side, though there
is an additional option for each to add these as the fallback RA field
to ensure others see them for the user's characters, regardless of their
addon or settings.