(1/2) Handle versions a bit better.
This is part one of a spread-out change. First up, handling versions a
bit better and trying to guess when someone has had their data wiped,
and discard all known versions when that's probably happened. Do this by
keeping an eye on the TT version -- if it's lower than we have cached,
it probably means they were wiped.