Remove unnecessary 'or nil' in many places.
Where the data isn't being saved or explicitly checked later as being
nil, just use whatever false/nil value we happened to end up with
instead of an explicit 'or nil'.

'or nil' is retained where nil is the only acceptable value, including
data which will be saved later (as 'false' would be saved, but 'nil'
just leaves it absent).