Make use of UnitIsMercenary()
This function was added to the Lua API late in the PTR cycle, but
provides a determination of if API functions and other parts of the game
will treat the unit as a mercenary. It *always* returns false when out
of visible range, but this coincides with the return from
UnitFactionGroup().

Occasionally, UnitFactionGroup() will also need to be checked due to the
range issues, but UnitIsMercenary() is often enough.