Fix UIDropDownMenu/Interface Options tainting
This adds a new, embeddable piece of code to fix the UIDropDownMenu
tainting issues from any/all addons, as far as I've been able to
determine, as well as the issue with tainting the Interface Options
panel. Making normal use of the dropdown menus or the addon section of
Interface Options should, when running this fix, never taint the UI.

The primary method of testing the tainting is through the infamous
issues with the built-in compact unit frames profiles interface options
panel, but the fixes should be generalized enough to stop the tainting
from ever occurring elsewhere, too.

In addition, the old warning code has been stripped out, as it no longer
applies with the new taint-preventing, rather than just taint-warning.