dropfix: Bad code fixes/issecurevariable()
This fixes interaction with bad code that creates tables with a
metatable either indexing a frame or using the frame metatable. Such
tables look like frames to a quick glance, but produce Lua errors
whenever a method is used. It's a terrible practice, but it's used in
some popular addons (i.e., MogIt's embedded Libra libary).

Additionally, make use of issecurevariable() when checking the keys of
the drop down buttons. It's a bit more overhead, but it's even less
likely to ever cause problems -- instead of stripping all non-default
keys, it strips all non-default keys which are insecure.