Adjust UI elements in editor/viewer
The primary goal of this change is to make the Notes feature more
visible and easier to access. The notes are now accessed, in the
viewer/editor, through a small, dedicated button that pops out the
notes panel.

In addition, the drop down menu in the editor has been removed. The main
profile management functions have returned to being buttons, and the
access to automation/notes are done by small buttons with icons rather
than text.

The help overlays have also been updated to account for this change.