BNMSP: Use constants and handle failure better
There are constants provided in _G for the game names returned by the BN
functions, and should be used. Also, with the failure earlier this year
to provide realm names, handle that sort of situation more gracefully in
the future (retrying a few times before giving up, rather than retrying
on every query).