Add friends-only mode for viewing profiles
This mode was requested by a user. It does not send requests for
profiles to anyone not on the friends lists (in-game and Battle.net) or,
optionally, in the current character's guild. It also does not permit
reception of unrequested field contents -- no addon does this by
default, but it's possible in the spec.