Add support for 6.2.4 Battle.net changes.
Toons were renamed to GameAccounts, along with a pile of functions like
that. Furthermore, presenceIDs are now broken into bnetIDAccount and
bnetIDGameAccount. This adds support for those changes, but retains
backwards compatibility with 6.2.3 (as, thankfully, the old and new
functions generally take the same arguments and return the same values).

The functions are also stricter about arguments, but XRP was already
using presenceIDs in a strict way.