Swap replacements off, Forsaken on by default
This changes the default settings of the %xt/%xf replacements to off by
default, as it consumes significant CPU time while typing and is not
widely used.

This also swaps the Forsaken replacement for Undead to on by default, as
Blizzard has been drifting more and more towards using Forsaken
themselves (i.e., Nemesis: Forsaken).