(INCOMPLETE) Add buttons to re-order glances
This is INCOMPLETE as-is. The functionality works, but the visuals need to be changed (currently shows three "X" buttons, rather than some sort of arrows on the up/down buttons).

As I'm unlikely to ever finish this work due to quitting WoW, this is presented in its incomplete form if anyone ever takes it up.